Update readme file (#2)
* Update readme.md

* Update readme.md

Readme update and added example use.

* Readme files and reorganization

Created some readme files for the file handling macros and organized them in folders. Also deleted the repeated version of batch analysis

* Update readme.md

* Update readme.md

* Update readme.md

* Version 2.0 and new files

Added some features: new interface, creation of a settings summary, works with a variable num of channels.
Also, added a readme file, example files and dataset.

* Update readme.md

* Output adjust and batch mode

Improvement of the saving format by keeping the hierarchy of the input folders and add the batch mode option, which allows you to use the macro in batch mode to accelerate it.